requirements.txt
streamlit
pandas
matplotlib
seaborn
numpy
google-cloud-bigquery
db-dtypes
pyarrow